more delete

+++ admin/delete_category.php
@@ -1,12 +1,20 @@
 <?php
 require_once('db.php');
-$id=$_GET['id'];
+if(isset($_GET['id']) AND isset($_GET['image_name'])){
+  $id=$_GET['id'];
+  $image_name=$_GET['image_name'];
 
-$sql="DELETE FROM category WHERE id=$id";
-$res=mysqli_query($conn,$sql);
-if($res==true){
-  $_SESSION['delete']="<div class='errors'> Category Deleted Successfully</div>";
-  header('location:manage_category.php');
+  
+  $sql="DELETE FROM category WHERE id=$id";
+  $res=mysqli_query($conn,$sql);
+  if($res==true){
+    $_SESSION['delete']="<div class='errors'> Category Deleted Successfully</div>";
+    header('location:manage_category.php');
+  }else{
+  die('errors');
+  }
 }else{
-die('errors');
+  header('location:manage_category.php');
 }
+
+
